tests/wpt/web-platform-tests/trusted-types/Element-outerHTML.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<div id="container"></div>
<script>
  var container = document.querySelector('#container')

  async_test(t => {
    createHTML_policy(window)
      .then(t.step_func_done(p => {
        let html = p.createHTML(INPUTS.HTML);

        var d = document.createElement('div');
        document.querySelector('#container').appendChild(d);
        d.outerHTML = html;
        assert_equals(container.innerText, RESULTS.HTML);

        while (container.firstChild)
          container.firstChild.remove();
    }));
  }, "outerHTML with html assigned via policy (successful HTML transformation).");
</script>
